Modularize status page
